+++ model_fetch/lib/model_fetch.dart
@@ -1,7 +1,11 @@
 export 'src/abstract_client_filter.dart';
 export 'src/abstract_filter.dart';
 export 'src/collection/collection_bloc.dart';
+export 'src/collection/empty_lazy_load_bloc.dart';
+export 'src/collection/iterable_lazy_load_bloc.dart';
 export 'src/collection/lazy_load_bloc.dart';
 export 'src/load_status.dart';
+export 'src/loader_factory.dart';
 export 'src/model_by_filter/model_by_filter_bloc.dart';
+export 'src/model_by_id/fixed_model_by_id_bloc.dart';
 export 'src/model_by_id/model_by_id_bloc.dart';