+++ model_fetch/lib/src/interfaces/async_disposable.dart
@@ -1,0 +1,3 @@
+abstract class AsyncDisposable {
+  Future<void> dispose();
+}