+++ model_fetch_firestore/lib/src/collection/firestore_lazy_load_bloc.dart
@@ -21,6 +21,7 @@ abstract class FirestoreLazyLoadBloc<T> extends LazyLoadBloc<T> {
   }
 
   @protected
+  // ignore: use_setters_to_change_properties
   void setLastDocument(DocumentSnapshot<T>? doc) {
     _lastDocument = doc;
   }