+++ model_fetch_firestore/lib/src/model_by_id/firestore_live_by_id_bloc.dart
@@ -26,4 +26,9 @@ class FirestoreLiveByIdBloc<T extends WithId<String>>
       ),
     );
   }
+
+  @override
+  void reload() {
+    // No-op, the model is live.
+  }
 }