+++ model_fetch_firestore_json_files/lib/src/file_iterable_builder.dart
@@ -1,0 +1,5 @@
+import 'dart:io';
+
+abstract class FileIterableBuilder {
+  Iterable<File> get files;
+}